refactor: refactor processRequest() loop
Break up the if-tree that checks for a Request's current process stage.
Abstractify each stage into their own class, under a superclass
RequestState. Each class will internally handle their part of parsing
the client message.
