+++ include/Client.hpp
@@ -6,7 +6,7 @@
 /*   By: cteoh <cteoh@student.42kl.edu.my>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/01 09:13:38 by kecheong          #+#    #+#             */
-/*   Updated: 2025/03/28 17:18:11 by cteoh            ###   ########.fr       */
+/*   Updated: 2025/03/30 02:53:33 by cteoh            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,6 +41,13 @@ public:
 	const Socket*		receivedBy;
 	Server*				server;
 	String				message;
+	char				timer;
+
+	enum Timeout {
+		KEEP_ALIVE = 0x001,
+		CLIENT_HEADER = 0x002,
+		CLIENT_BODY = 0x004
+	};
 private:
 	std::vector<char>		messageBuffer;
 	static const size_t		MAX_REQUEST_SIZE;
@@ -48,16 +55,9 @@ private:
 	std::deque<Request>		requestQueue;
 	std::deque<Response>	responseQueue;
 
-	char					timer;
 	std::time_t				keepAliveTimeout;
 	std::time_t				clientHeaderTimeout;
 	std::time_t				clientBodyTimeout;
-
-	enum Timeout {
-		KEEP_ALIVE = 0x001,
-		CLIENT_HEADER = 0x002,
-		CLIENT_BODY = 0x004
-	};
 };
 
 #endif
