+++ src/message/request/RequestState.cpp
@@ -1,0 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   RequestState.cpp                                   :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: cteoh <cteoh@student.42kl.edu.my>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/03/29 21:45:30 by cteoh             #+#    #+#             */
+/*   Updated: 2025/03/31 08:11:43 by cteoh            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "RequestState.hpp"
+
+RequestState::~RequestState(void) {}
